requirements.txt

attrs==22.1.0
aws-cdk-lib
constructs
aws_cdk.aws_apigatewayv2_integrations_alpha
black==22.10.0
boto3==1.26.25
click==8.1.3
pre-commit==2.20.0
PyYAML==6.0
requests==2.28.1
pytest==7.2.0
pytest-dotenv==0.5.2
moto==4.1.7
requests==2.28.1
coverage==7.2.3